Codechecker does not recognize unneeded doc for overriden methods.
It is good practice to not re-write the documentation for overriden
methods, e.g. question_type::get_question_options etcetera.
But this yields a warning in codechecker since that cannot detect
(yet) that it is such an overridden message. I've seen an issue opened
about this somewhere. But this commit is just to ignore the warnings
and let the CI tests pass silently.